% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_pmc_to_pmid}
\alias{convert_pmc_to_pmid}
\title{Convert PMC IDs to PMIDs}
\usage{
convert_pmc_to_pmid(pmc_ids)
}
\arguments{
\item{pmc_ids}{Vector of PMC IDs (with or without "PMC" prefix)}
}
\value{
Vector of PMIDs (NA for failed conversions)
}
\description{
Uses NCBI E-utilities to convert PMC IDs to PubMed IDs
}
\keyword{internal}
